+++ src/App.css
@@ -13,14 +13,6 @@
   padding: 0 5rem;
 }
 
-.App-footer{
-  margin-top: 5rem;
-  bottom: 0;
-  left: 0;
-  background-color: #000000;
-  height: 4rem;
-}
-
 .dropdown-menu.show {
   right: 0!important;
   left: auto!important;
