+++ src/index.css
@@ -1,4 +1,11 @@
+html,
 body {
+  height: 100%;
+  font-family: sans-serif;
+}
+
+body {
+  padding: 0;
   margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
     'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
@@ -10,4 +17,26 @@ body {
 code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
+}
+
+#root {
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+}
+
+.App {
+  flex: 1 0 auto;
+}
+
+.Footer {
+  flex-shrink: 0;
+  text-align: center;
+  background-color: #000000;
+  padding-top: 20px;
+}
+
+.Footer ul {
+  list-style-type: none;
+  padding-left: 0!important;
 }
