+++ src/index.js
@@ -8,6 +8,7 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 import 'font-awesome/css/font-awesome.min.css';
 
 import App from './App';
+import Footer from './Footer';
 import * as serviceWorker from './serviceWorker';
 import reducers from './redux/reducers/index';
 import './index.css';
@@ -21,7 +22,8 @@ export const store = createStore(
 
 ReactDOM.render(
   <Provider store={store}>
-    <App />
+    <App key="1" />
+    <Footer key="2" />
   </Provider>,
   document.getElementById('root')
 );
